+++ src/main/java/terramine/common/init/ModSoundEvents.java
@@ -21,6 +21,7 @@ public class ModSoundEvents {
 	public static final SoundEvent FLAMELASH_SHOOT = register("item.flamelash.use");
 	public static final SoundEvent RAINBOW_ROD_SHOOT = register("item.rainbowrod.use");
 	public static final SoundEvent SPACE_GUN_SHOOT = register("item.spacegun.use");
+	public static final SoundEvent BOSS_ROAR_0 = register("entity.boss.boss_roar_0");
 
 	// Accessories
 	public static final SoundEvent FART = register("item.whoopee_cushion.fart");