+++ src/main/java/terramine/common/item/misc/TreasureBagItem.java
@@ -1,0 +1,65 @@
+package terramine.common.item.misc;
+
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.minecraft.advancements.CriteriaTriggers;
+import net.minecraft.network.chat.Component;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResultHolder;
+import net.minecraft.world.entity.*;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.item.UseAnim;
+import net.minecraft.world.level.Level;
+import org.jetbrains.annotations.NotNull;
+import terramine.TerraMine;
+import terramine.common.item.TerrariaItemConfigurable;
+
+import java.util.List;
+
+// todo: make sure is client sided only, only local player can see and pickup the item
+public class TreasureBagItem extends TerrariaItemConfigurable {
+    public TreasureBagItem(Properties properties) {
+        super(properties);
+    }
+
+    @Override
+    public ItemStack finishUsingItem(@NotNull ItemStack itemStack, @NotNull Level level, @NotNull LivingEntity entity) {
+        Player player = (Player) entity;
+        if (player instanceof ServerPlayer) {
+            CriteriaTriggers.CONSUME_ITEM.trigger((ServerPlayer)player, itemStack);
+        }
+
+        if (!level.isClientSide) {
+            // Open custom inventory with a few slots filled with items from the bosses loot table
+            if (!player.getAbilities().instabuild) {
+                itemStack.shrink(1);
+            }
+        }
+        return super.finishUsingItem(itemStack, level, entity);
+    }
+
+    @Override
+    public UseAnim getUseAnimation(@NotNull ItemStack stack) {
+        return UseAnim.BLOCK;
+    }
+
+    @Override
+    public InteractionResultHolder<ItemStack> use(@NotNull Level world, Player user, @NotNull InteractionHand hand) {
+        ItemStack itemstack = user.getItemInHand(hand);
+        user.startUsingItem(hand);
+        return InteractionResultHolder.consume(itemstack);
+    }
+
+    @Override
+    public int getUseDuration(@NotNull ItemStack stack) {
+        return 0;
+    }
+
+    @Override
+    @Environment(EnvType.CLIENT)
+    public void appendHoverText(@NotNull ItemStack stack, Level world, @NotNull List<Component> tooltip, @NotNull TooltipFlag flags) {
+    }
+}